% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-packages.R
\name{build_ropensci_packages_json}
\alias{build_ropensci_packages_json}
\title{Build rOpenSci packages.json}
\usage{
build_ropensci_packages_json(out_file = "packages.json")
}
\arguments{
\item{out_file}{where to save the JSON file}
}
\description{
Build rOpenSci packages.json
}
