% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemeta.R
\name{create_codemetas}
\alias{create_codemetas}
\title{Create the codemetas for all files}
\usage{
create_codemetas(old_cm = NULL, folder = "repos")
}
\arguments{
\item{old_cm}{path to latest CodeMeta version}

\item{folder}{folder under which the}
}
\value{
A JSON codemeta
}
\description{
Create the codemetas for all files
}
