% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{create_registry}
\alias{create_registry}
\title{Title}
\usage{
create_registry(cm, outpat, time = Sys.time(), folder = "repos")
}
\arguments{
\item{cm}{Path to the JSON codemeta}

\item{outpat}{Path where to save the JSON}

\item{time}{Time to add at the end}
}
\description{
Title
}
