% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-versions.R
\name{registry_pkg_versions}
\alias{registry_pkg_versions}
\title{Ensure CRAN versions are released on GitHUb}
\usage{
registry_pkg_versions(CRAN_only = TRUE)
}
\description{
Ensure CRAN versions are released on GitHUb
}
