% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-repos.R
\name{track_repos}
\alias{track_repos}
\title{Clone all ropensci packages & create `"last_commits.csv"`}
\usage{
track_repos(
 
    pkgs_file = "https://raw.githubusercontent.com/ropensci/roregistry/gh-pages/packages.json"
)
}
\arguments{
\item{pkgs_file}{`"packages.json"` file for all rOpenSci packages.}
}
\description{
Clone all ropensci packages & create `"last_commits.csv"`
}
